+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - clang-cl is now supported on Windows
 
 ### Fixed
+- `statement::bind(int, std::unique_ptr<T>)` now actually compiles
+- `value::get<T *>` now actually compiles
 - The library headers are now compiling cleanly with `-Wconversion -Wsign-conversion` on clang
 
 ## [1.7] - 2026-05-26